Added response to results page
