+++ flight-search-fe/src/pages/ResultsPage.tsx
@@ -1,13 +1,148 @@
 import { useSelector } from "react-redux";
 import type { RootState } from "../redux/store";
+import {
+  Box,
+  Card,
+  CardContent,
+  Typography,
+  Button,
+  Paper,
+} from "@mui/material";
+import FlightTakeoffIcon from "@mui/icons-material/FlightTakeoff";
+import ReturnToSearchButton from "../components/ReturnToSearchButton";
+import Airplane from "../assets/AirplaneBackground.jpg";
+import "../App.css";
 
 export default function ResultsPage() {
-  const flights = useSelector((state: RootState) => state.searchResults.results);
+  const flights = useSelector(
+    (state: RootState) => state.searchResults.results
+  );
 
   return (
-    <div>
-      <h2>Raw Flight Data</h2>
-      <pre>{JSON.stringify(flights, null, 2)}</pre>
-    </div>
+    <Box
+      sx={{ padding: "0 100px", backgroundColor: "rgba(74, 131, 229, 0.556)" }}
+    >
+      <Paper
+        sx={{
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "center",
+          backgroundImage: `url(${Airplane})`,
+          backgroundSize: "100%",
+          backgroundPosition: "-175% center",
+          backgroundColor: "rgba(210, 225, 255, 0.83)",
+          backgroundBlendMode: "soft-light",
+          width: "100%",
+        }}
+      >
+        <div style={{ width: "100%", marginBottom: "15px" }}>
+          <Box
+            sx={{
+              position: "relative",
+              display: "flex",
+              alignItems: "center",
+              borderBottom: "1px solid rgb(227, 227, 228)",
+              borderRadius: "5px",
+              backgroundColor: "white",
+            }}
+          >
+            <FlightTakeoffIcon
+              sx={{
+                fontSize: 25,
+                color: "rgb(56, 75, 218)",
+                mb: 1,
+                ml: 1,
+                mr: 1,
+              }}
+            />
+            <h3 style={{ fontWeight: "500" }}>Flight Search</h3>
+          </Box>
+        </div>
+        <div
+          style={{
+            width: "100%",
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "center",
+          }}
+        >
+          <Box sx={{ width: "65%", mb: 4, mt: 2 }}>
+            <ReturnToSearchButton />
+          </Box>
+          <Typography
+            variant="h5"
+            gutterBottom
+            sx={{ fontWeight: "500", width: "100%", fontSize: "22px" }}
+          >
+            Departing Flights
+          </Typography>
+          {flights.length === 0 ? (
+            <p>No flights found.</p>
+          ) : (
+            <ul style={{ width: "100%" }}>
+              {flights.map((flight, index) => (
+                <Card key={index} sx={{ mb: 4, width: "65%" }}>
+                  <CardContent
+                    sx={{
+                      display: "flex",
+                      justifyContent: "space-between",
+                      alignItems: "center",
+                    }}
+                  >
+                    <Box
+                      sx={{
+                        display: "flex",
+                        flexDirection: "column",
+                        alignItems: "center",
+                      }}
+                    >
+                      <Typography sx={{ fontWeight: "600", fontSize: "18px" }}>
+                        {flight.departure} - {flight.arrival}
+                      </Typography>
+                      <Typography>{flight.airline}</Typography>
+                      <Button
+                        variant="outlined"
+                        size="medium"
+                        sx={{ mt: 1 }}
+                        onClick={() => console.log("Go to details", flight.id)}
+                      >
+                        View Details
+                      </Button>
+                    </Box>
+                    <Box>
+                      {" "}
+                      <Typography sx={{ fontWeight: "600", fontSize: "16px" }}>
+                        {flight.from} → {flight.to} ({flight.formattedDuration})
+                      </Typography>
+                    </Box>
+                    <Box>
+                      <Typography sx={{ fontWeight: "600", fontSize: "16px" }}>
+                        Stops: {flight.numberOfStops} ({flight.stops})
+                      </Typography>
+                    </Box>
+                    <Box>
+                      <Typography sx={{ fontWeight: 600 }}>
+                        {flight.currency} ${flight.price}
+                        <Typography sx={{ display: "inline" }}>
+                          {" "}
+                          Total
+                        </Typography>
+                      </Typography>
+                      <Typography sx={{ fontWeight: 600 }}>
+                        {flight.currency} ${flight.price}
+                        <Typography sx={{ display: "inline" }}>
+                          {" "}
+                          Per Traveler
+                        </Typography>
+                      </Typography>
+                    </Box>
+                  </CardContent>
+                </Card>
+              ))}
+            </ul>
+          )}
+        </div>
+      </Paper>
+    </Box>
   );
 }
